started doing process

+++ lab_05/inc/arr_queue.h
@@ -12,7 +12,7 @@ typedef struct arr_queue
     int pout; // указатель на конец
     int size; // размер массива
     int count; // количество элементов в очереди
-    float *arr; // массив, элементы коротого - время обработки
+    double *arr; // массив, элементы коротого - время обработки
 } arr_queue;
 
 
@@ -20,10 +20,10 @@ typedef struct arr_queue
 int arr_queue_init(arr_queue *queue);
 
 // Добавление элемента в конец очереди
-int arr_queue_push(arr_queue *queue, float process_time);
+int arr_queue_push(arr_queue *queue, double process_time);
 
 // Удаление элемента из очереди
-int arr_queue_pop(arr_queue *queue, float *ret_time);
+int arr_queue_pop(arr_queue *queue, double *ret_time);
 
 // Освобождение памяти под очередь
 void arr_queue_free(arr_queue *queue);
