+++ lab_05/inc/list_queue.h
@@ -8,7 +8,7 @@
 */
 typedef struct list_node
 {
-    float data; // время обработки данной заявки
+    double data; // время обработки данной заявки
     struct list_node *next; // следующая заявка
 } list_node;
 
@@ -23,13 +23,13 @@ typedef struct list_queue
 void list_queue_init(list_queue *queue);
 
 // Создание ноды списка
-int create_node(list_node **node, float tmp);
+int create_node(list_node **node, double tmp);
 
 // Добавление элемента в очередь
-int list_queue_push(list_queue *queue, float tmp);
+int list_queue_push(list_queue *queue, double tmp);
 
 // Удаление элемента из очереди
-int list_queue_pop(list_queue *queue, float *tmp);
+int list_queue_pop(list_queue *queue, double *tmp);
 
 // Очистка очереди
 void list_queue_clean(list_queue *queue);
