+++ lab_05/inc/time_func.h
@@ -2,4 +2,4 @@
 #include <stdio.h>
 #include <time.h>
 
-float rtime(int left, int right);
+double rtime(int left, int right);
