+++ lab_05/src/time_func.c
@@ -1,6 +1,6 @@
 #include "../inc/time_func.h"
 
-float rtime(int left, int right)
+double rtime(int left, int right)
 {
-    return (right - left) * ((float)rand() / (float)RAND_MAX) + left;
+    return (right - left) * ((double)rand() / (double)RAND_MAX) + left;
 }
